requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn==0.24.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Environment and Configuration
python-dotenv==1.0.0

# Data Validation
pydantic==2.5.1
pydantic[email]

# Testing (optional but recommended)
pytest==7.4.3
httpx==0.25.2  # For async testing
pytest-asyncio==0.21.1

# Development Tools
black==23.11.0  # Code formatting
flake8==6.1.0   # Code linting
isort==5.12.0   # Import sorting